Record: 11L Tight SWA + Partial RoPE + LN Scale + XSA4 (val_bpb: 1.1270)
3-seed mean sliding BPB: 1.1270 (std 0.0006)
Best seed (42): 1.1265
Artifact: 15.5MB, 8xH100 SXM, 600s

Key technique: Tight SWA (scale<0.2, every 50 steps) - zero SWA penalty.

Made-with: Cursor
