--- records/track_10min_16mb/2026-03-21_11L_XSA_PartialRoPE_EMA_QAT_TTT/submission.json
@@ -1,13 +1,0 @@
-{
-  "author": "sadegh.shokatian",
-  "github_id": "sadegh-shokatian",
-  "name": "11L + XSA + PartialRoPE + LN Scale + EMA + Late QAT + TTT + TrigramHash + Adaptive Quant",
-  "blurb": "11 layers with XSA on last 4 layers, partial RoPE (16/64 dims), LN scale damping, EMA (0.997), late QAT with STE int6/int5 fake-quantization, backward-looking TTT (3-epoch SGD), TrigramHash embedding, gradient-guided adaptive int5/int6/int7 quantization, sliding window eval stride=32, SmearGate + BigramHash(10240) + OrthoInit + Muon WD=0.04 + zstd-22.",
-  "date": "2026-03-21",
-  "val_bpb": null,
-  "val_loss": null,
-  "bytes_total": null,
-  "bytes_model": null,
-  "bytes_code": null,
-  "seeds": [1337, 42, 7]
-}
